DWA.Core.Dir: added root_dir, etc, moved from .State
That's a very generic part of the code, State stores much more,
command registry and others. These modules are imported by State
but if any of these modules needs get_dir(), for instance, there
would be a circular reference of the modules. This is the only
and clean way to resolve it.

DWA.Core.{Tests,Main} also uses it.