Introducting Plugins, Plugin Registry
This is the initial part of the whole feature set, the planned behaviour
is in doc/plugins.txt.

The PluginRegistry supports different plugins including the core itself,
and the load_plugins() makes the commands to be available. As result,
the different Commands modules' __init__.py can be empty, such as
in case of the core DWA commands (DWA.Commands module).

Samples are added.

The functionality is not yet finished, only one type of command modules
can be registered: file format: A/B/C.py, where
  - A/B or A.B is the commands module
  - C.py stores CCommand class which can be called as "c"

It's slower at this point as registering commands via __init__.py, because
DWA searches for commands at every start. In the future there will be
some kind of caching mechanism.

Additionally the tests of the plugin registry don't cover everything,
especially not the loading of the plugins.