plugins: implemented with caching, commands / toolscommands supported
The plugin framework is implemented, but unit/functional tests are missing.
How it works: see doc/plugins.txt

The plugin registry's load_plugins() method loads the plugins, and
creates cache files to fast startup. As a result, when the available
commands' number is high, it becomes much more faster because less files
are read.

The cache files are in lib/DWA/Var/. The update will work by the newly
added (in next patch) selfupdate command.

Note that tools commands are not yet functional.