core: added Context
Context is for storing entries which can be easily accessible, it's
some kind of registry. It's for avoid such codes which was introduced
in commit 9c10f35fce2315abd63e6157b5ea7030780c063b (DWA.Core.State),
because this context object can be passed to functions, and avoids
global state.