Added command-line tool and wire in basic argument parsing
From now on DEWI is usable from command line, in the following way:
 dewi-local -p dewi.commands.sample.SamplePlugin sample

dewi-local is simply a wrapper around the dewi module, and it is not needed
when DEWI is installed from pip or something like that.

The -p / --plugin option is for registering plugins. By default the new
DewiPlugin is registered, but this can be replaced via command-line args.
The first argument is the command to be run.