MainApplication: add '--wait' parameter to wait before exiting
If the process is started in a window, the output of the application is
visible for a moment, then it is lost due to the closed window. Because
of this, it is necessary to wait for some time before exiting.

The --wait option is for waiting application termination by reading one
input line (in fact, for ENTER).