dewi.module_framework: introduce another plugin system
What if...
If a command needs to have plugins, it can be done, but dependency needs
to be handled and so on. These plugins might be the same as the
current plugin system, but it's more likely that it is used only for
separation, and doesn't need to be loaded with the PluginLoader.

This is "Module". It's main purpose is to process some data,
store the results in Config, and add user-visible messages to Messages.