added logparser module to process logs and collect messages
Sometimes it is handy to search specific messages and create some kind
of message, except classify the logs. In the modules subpackage two
example parser module is added for reference.

The LogHandlerModule is an example for the previously added
dewi.module_framework, too.

The logparser.syslog.Parser() class assumes specifically formatted logs,
the date and time is formatted as DATE_ATOM in PHP:
2005-08-15T15:52:01+00:00
everything else is the widely used.