logparser: fixed example modules + add two new example
Since the callback function parameter list is changed, the examples
were broken. These are fixed.

Two basic modules are also added, one to track reboot, and the other
one to track blocked processes.

All of the examples are working, might be useful in real-life scenarios.
Except the "ha" module, because as of now I don't know how could it be
used, but it was the first module I wrote to test the framework.