commands: add change dir from Windows 10 to Ubuntu on Windows (same dir)
I needed a quick tool that changes the environment from Windows 10
to Bash on Ubuntu On Windows, but keeping the current directory.

I call it "cu" "(change 2 ubuntu) and "chroot", because it is something
like that. Perhaps not the best names, but it might work.

It assumes that the SSH service is started in Ubuntu, and it's accessable
on 127.0.0.1:22.