dewi.utils.dictionaries: new module, that wraps special dict types
The DictionaryWithList and DictionaryWithSet type are for easier handling
of any dict() which has list() or set() members.

In other words, it wraps the code that checks whether a key exists, or not,
and if it is missing, adds a set or list object, otherwise appends to it.

The usage is tricky, I decided to add this to __setitem__.

E.g.:
>>> import dewi.utils.dictionaries as dm
<class 'dewi.utils.dictionaries.DictionaryWithList'>
>>> d = dm.DictionaryWithList()
>>> d
{}
>>> d['a_key'] = 'first'
>>> d
{'a_key': ['first']}
>>> d['a_key'] = 'second'
>>> d
{'a_key': ['first', 'second']}