dewi/utils: add process, find_binary/which to search for executable in PATH
The PATH variable contains directories that can contain executable
programs, or binaries. The find_binary() method searches a binary in
these directories.

On Windows, the binary name may be with or without extension, and if
it is without it, the PATHEXT variable contains the list of extensions
to be append when the binary lookup occurs.

I saw a more basic implementation in Gentoo's portage, and I decided
to use the same name, because it's more intuitive name than the shell's
name for it, 'which', which is good for shell, but can be misleading
in a Python software.

Despite this, I also added a wrapper method to this lookup, called
as the shell's name, 'which'.