dewi/utils: add xml module to convert an XML document to a dict
While it seems that it is a good idea to convert an XML document to a dict,
it has some limitations, especially when an element is repeated
unexpectedly, therefore in the result some of the entries might be lists,
if an element is repeated, otherwise the content of the element is used
there. In all other cases the behavior is straightforward.