dewi.realtime_sync: a framework to sync a code (files) to a remote location - file data
dewi.realtime_sync is a framework for synchronizing code or other files
to a remote location, just like rsync does.

Currently some files of the framework are added for indicating intentions
of further actual codes and classes, and the filesync_data module.

The filesync_data module is basically finished already, which describes
how to sync actually changed files. It is just for the config part
of the sync - the mapping between the local and remote names, including
its management, the lookup, and the calculation of the remote name.

All file names are relative, the filesync_data module doesn't handle the
root directories of the source and target locations. Further classes
will be added with this responsibility.