dewi.realtime_sync: add filesystem package, to copy / chmod / etc. as needed
The different Filesystem classes are responsible to do the different steps
of the synchronization:
* chown
* chmod
* makedir
* remove
and helpers:
* exists
* is_dir

RemoteFilesystem is to do these using SSH
LocalFilesystem is for local sync, on the same machine

Both of these classes are for syncing local source code to its testing
location.

The Filesystem defines the interface, and CommandTrackerFilesystem
might be useful for auditing, but it's certainly good for testing
purposes (without exists() and is_dir()).