realtime_sync.syncers: fix a bug that remote full path did not contain separator
The calculated remote full name needs to contain a path separator between
the remote directory name and the remote relative path. It's fixed.

Originally the test used absolute path for the remote (relative) path, this
bug was hidden, because the DetailedEntry's remote name was not calculated,
but injected into the test. This was revealed when the code became fully
functional - the remaining parts will be added in other commits.

The code was working because of the absoulte path, and with os.path.join
instead of '+' it would fail - in syncers.py now os.path.join is used.

The test now uses relative path in remote name, and uses manually typed
paths instead of calculated (eg. by os.path.dirname), to spot any other
issue.