new command: collect images - to collect photo (image) files into sqlite
The command creates an SQLite file if necessary, and store the original
path, the file size and modification date, and also MD5 checksum.
Furthermore it stores an uppercase form of the file's base name.
Only .jpg / .jpeg / .cr2 / .mov / .thm files are collected, the lower or
upper case of the filename does not matter.