module fw's module: store messages in config with (sub)category keys
Origianlly the messages were appended to 'messages' key (or custom key).
Now it is changed to use categories and subcategories, therefore
the messages are appended to
<messages-key>.<category>.<subcategory>

The change is useful for representation, it's easier to read by a human
and by visualization using (sub)categories, such as an HTML output.