dewi.rrdtool: add possibility to modify loaded config
A datafile may contain too generic configuration, and some parts are not
needed or not needed in the original way.
An example: the loop0...loop9 (loopX) devices may not be needed in the
final result.

For this a ConfigModifier type is introduced in the modifier.py.
This is a superclass, the default modifiy() method simply iterates through
the plugins and call modify_plugin.