rrdtool/rrdtool: add reference_datetime to use TZINFO for calculated end time
If no end time is given, the end time is read from the RRD file.
The RRD file does not contain timezone information, only a UNIX timestamp.
To convert it to a full datetime object it's useful to give timezone
information. This can be achivieved most easily by passing a datetime
object as a reference.