utils/kayako_rest: add Connection cls and ensure_connection() method
The previously existing Request class is good for one-shot HTTP requests,
but usually an application uses multiple request, therefore it is better
to use an object for every possible request.
This is Connection.

As a result the calculations can be done once, and reuse it multiple times,
which makes the code a little bit faster.
The object creation before calling fetch(), post() or put() is not
necessary anymore, which makes the code cleaner.