logparser: enforce '\n' as line separator
In case of systemd-nspawn the progress of the job is logged into a single
line. The console might contain something like this:
[*   ] A start job is running for ...
[**  ] A start ...
[ ** ] A start job...

It's logged in a single line with delimiter '\r', which is a line separator
by default. To prevent split of this log line, '\n' is added to open()
as an explicit line separator.