core: Context always contains 'commands' and 'commandregistry', added props
Added properties that wraps 'commands' and 'commandregistry' entries,
'commands' and 'command_registry' respectively.
Context.__init__ creates these objects.

The CommandRegistryPlugin is not needed anymore, as its functionality
is merged into the Context object.

Wherever CommandRegistryPlugin was loaded, the CorePlugin is loaded instead.
CorePlugin itself is an empty plugin, it's kept for future compatibility.