dewi.core: add logger module, and add options to MainApplication
--log-none, --log-console, --log-file and log--none are added to the app.
The log level can be set by the --log-level option.

Multiple targets can be specified, eg. --log-file a.txt --log-file b.txt
--log-syslog.
The --log-none cannot be combined with others.
The default is --log-console.

The dewi.core.logger module contains the logging code,
the module's log_* method can be used for logging (eg. log_debug).

The first log param after message can be a dict, kwargs is updated with it,
as the following usage doesn't work in certain cases:
log_...("message", key="value", ...)
where the key is eg. class, because it is a reserved word.

Solution:
log_...("message", {'class': '....'})