packt command: download my ebooks from PacktPub / basic selenium scripts
This command is a lightweight selenium-related library of another one,
therefore it seems to be a bit larger than it should be.
The code is prepared for Firefox, but it uses Chrome as of now.
The runner also could support multiple subcommands, but not implemented
here.

It supports dismissing the cookie dialog.
The config file is ~/.config/dewi/packt.conf
It has two lines:
<email>
<password>

~/.config/dewi/chromedriver should point to the actual chromedriver

The download directory must exist but must be empty before the first run,
as the command deletes all previous files, but keeps directories.
The ebooks and source codes will be in the
 <download-dir>/ebook/<title>
directory

If the server provides, a cover.png is also stored, may be 0 if
the server responses with 500 Internal Server Error :(
The price.txt contains the price of the ebook.

Usage:
./dewi-local.py-p dewi.commands.packt.PacktPlugin packt \
   --screenshots ~/packt-screenshots \
   --download-directory ~/packt-dl