Replace __ prefix with _ as it's beter practice, more common and easier to debug
Common libraries barely use __ to hide members. Python has no real private
or protected visibility, the only relevant is whether the first character
is '_' or not. If it starts with '_', it should not be called from outside,
except from subclasses.

I left one __ prefix in Node class, as it can have subclasses and they
should not override that method.