lock: a simple file lock context manager, and wrapper lock_directory
FileLock can ensure locking of a file between processes.
The implementation is simple, it cannot time out, therefore it needs
to be used carefully.
It is also a context manager for the "with ...:" statement.

A related functionality is to lock an entire directory, for that purpose
a ".file_lock_" file is created in that directory using FileLock.
The (wrapper) method is lock_directory().