rrdtool: make writer capable of parallel run
The graph generation is time consuming, and it worth to run parallelly.
In order to do that both the wrapper RrdTool class and the GraphWriter
class got the parallel_run_count parameter with default value 1,
which results the original behaviour.

The recommended value is:
max(1, multiprocessing.cpu_count() - 1)

In case of a dedicated VM the multiprocessing.cpu_count() is also fine.