add 'files' module with find_file_recursively() to run search like '.gitignore' lookup
Sometimes it is useful to look up a file in similar way as 'git' searches
'.gitignore' files, and the '.git/' directory: recursively searches in
parent directories, too.
The find_file_recursively() does the same, searches a specific filename
like '.gitignore' in a specified directory, or if it is not specified,
in the current directory and then their parent directories.