add threading module for jobs/pool where a job may trigger several others
The use case is:
- search for git commits for specific Jira tickets
- then search for the branches containing the corresponding commits

The git command is slow and the machines have multiple cores, so it's
better to use multiple threads.

The second step can also be a separate job, especially in the following
case:
- PROJECT-1 ticket has 2 commits
- PROJECT-2 has 15 commits
- PROJECT-3 has only 1
and the CPU / thread count is at least 8 (or 4). The result can be
acquired faster.

The state member of pool is to store the results (and the state object
must guarantee the thread safety).