threading: support map-reduce in jobs
Any job can trigger one or more jobs after it is completed, like in
the previous commit. But it also can define a single job that runs
after each started additional jobs are completed.

The first list of jobs can work as mappers, and the last is as a reducer.
It could also be called as a postprocess job.